# src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: DreamSellers API
  description: |-
    This is the API specification for the DreamSellers platform, a web application for booking experience-based gifts. Users can browse, book, and manage various adventure and leisure activities, while businesses can list and manage their experience offerings.
    
    Some useful links:
    - [DreamSellers Platform](https://dreamsellers.com)
    - [API Documentation](https://api.dreamsellers.com/docs)
  termsOfService: https://dreamsellers.com/terms/
  contact:
    email: support@dreamsellers.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://api.dreamsellers.com
    description: Production server
  - url: https://staging.api.dreamsellers.com
    description: Staging server
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Login user into the system.
      description: Endpoint for handling logging in.
      operationId: userLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
        required: true
      responses:
        '200':
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUserDto'
        '400':
          description: Invalid username/password supplied
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Wrong password or login, try again"
        '401':
          description: Unauthorized, invalid username/password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Wrong password or login, try again"
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "To many requests, try again later"
        '500':
          description: Internal server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error processing the request"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/register:
    post:
      tags:
        - auth
      summary: Register user in the system
      description: Registers new users in the system
      operationId: userRegister
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUserDto'
        "409":
          description: Conflict - email or username already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "An account with this email already exists."
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "To many requests, try again later"
        '500':
          description: Internal server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error processing the request"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    get:
      tags:
        - auth
      summary: Logs out current logged in user session.
      description: Log user out of the system.
      operationId: logoutUser
      parameters: []
      responses:
        '500':
          description: Internal server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error processing the request"
        '200':
          description: successful operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    AuthenticatedUserDto:
      type: object
      properties:
        uid:
          type: integer
          example: 1
        token:
          type: string
          example: "ASDFGDFHDFgscdcsdGJTRFGWVCSwz....."
    RegisterDto:
      type: object
      properties:
        username:
          type: string
          example: "username"
        password:
          type: string
          example: "123"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        email:
          type: string
          example: "johndoe@example.com"
        phone:
          type: string
          example: "+1234567890"
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
            - ROLE_BUSINESS
          example: "ROLE_USER"
    LoginDto:
      type: object
      properties:
        username:
          type: string
          example: "username"
        password:
          type: string
          example: "password"
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            "write:pets": modify pets in your account
            "read:pets": read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header